remove agnostic text measurement back into text.rs
